changed query time subscriptions to a multiline string
ros2 params does not have a dictionary type. If we want to pass multiple
query time subscriptions, each with a topic and optional message field,
we need to pass this as a multiline string (still following the yaml
structure). A bit hackey, as you need to remember to add the pipe (|)
operator after query_time_subscriptions, but it works pretty well