added conversions for standard ros MSG types, which use the python builtin types conversions applied to the message .data field
for standard messages in ros, like Int, Bool, String, the `.data` field
of the message is comprised of a standard python datatatype.
Having already added standard python datatypes to the registry of ros
msg -> rdf literal conversions, the only thing to do is to pass the
`.data` field of the message through the registry to obtain the rdf
literal representation.